refactor: make gutenberg_custom_blocks a part of monorepo
This module had it own yarn.lock file. This was made to improve the Lagoon deployments speed. But now we don't need this:
1. We don't deploy to Lagoon anymore.
2. If we deploy to Lagoon again, we will most likely use the regular monorepo approach, so we will drop the monorepo dance (https://github.com/AmazeeLabs/silverback-mono/blob/e4019bbc9c379d4c15c65f99000d73f24752d94b/apps/silverback-gatsby/lagoon/gatsby-preview.dockerfile#L6).

Also, the change should help with the Renovate updates. Because sometimes Renovate updates all package.json files, but only the root yarn.lock. It looks like it does not expect to find more than one yarn.lock in a monorepo.